[Refactor] src > assets로 이미지 파일 정리 및 미사용 레이아웃 컴포넌트 각주 처리
