+++ src/assets/icons/logo/main-logo-icon-2.png
