+++ src/assets/icons/logo/main-logo-icon.png
